add null check

+++ CleanArchitecturalPattern.Application/Services/CustomerService.cs
@@ -27,8 +27,12 @@ namespace CleanArchitecturalPattern.Application.Services
         public void DeleteCustomer(Guid id)
         {
             var customer = _customerRepository.GetByIdAsync(id).GetAwaiter().GetResult();
-            _customerRepository.Delete(customer);
-            _customerRepository.SaveChanges();
+            if (customer != null)
+            {
+                _customerRepository.Delete(customer);
+                _customerRepository.SaveChanges();
+            }
+            
         }
     }
 }
